+++ scalaz/shared/src/main/scala/testz/Fold.scala
