+++ scalaz/shared/src/main/scala/testz/z.scala
@@ -124,4 +124,55 @@ object z {
     }
   }
 
+  object streaming {
+
+    def runTests[F[_]: Applicative, G[_]: Applicative, I](
+      testGenerator: I => F[Result]
+    ): Fold[G, I, F[Result]] =
+      new Fold[G, I, F[Result]] {
+        type S = List[F[Result]]
+        val start = Nil
+        def step(s: List[F[Result]], i: I): G[List[F[Result]]] =
+          (testGenerator(i) :: s).pure[G]
+        def end(s: List[F[Result]]): G[F[Result]] = {
+          s.foldRight(Succeed().point[F])(
+            Applicative[F].apply2(_, _)(Result.combine)
+          ).pure[G]
+        }
+      }
+
+    def exhaustive[F[_]: Applicative, I]
+                  (in: List[(() => I)])
+                  (testGenerator: I => F[Result])
+                  : F[Result] =
+      in.foldLeft(Succeed().point[F])((b, a) =>
+        Applicative[F].apply2(b, testGenerator(a()))(Result.combine)
+      )
+
+    def exhaustiveU[F[_]: BindRec, I]
+                  (in: Unfold[F, I])
+                  (testGenerator: I => F[Result])
+                  (implicit F: Monad[F]): F[Result] = {
+      F.join(exhaustiveUR[F, F, I](in)(testGenerator))
+    }
+
+    def exhaustiveUR[F[_]: Applicative, G[_]: Monad: BindRec, I]
+    (in: Unfold[G, I]
+    )(testGenerator: I => F[Result]
+    ): G[F[Result]] = {
+      zapFold(in, runTests[F, G, I](testGenerator))
+    }
+
+    def exhaustiveV[F[_]: Applicative, I]
+                  (in: (() => I)*)
+                  (testGenerator: I => F[Result]): F[Result] =
+      exhaustive(in.toList)(testGenerator)
+
+    def exhaustiveS[F[_]: Applicative, I]
+                  (in: I*)
+                  (testGenerator: I => F[Result]): F[Result] =
+      exhaustiveV(in.map(i => () => i): _*)(testGenerator)
+
+  }
+
 }
